Update price with add taxes, RAM choice, typo fix
What: Adds wording in the 'What you will get' and
'How much will this cost' sections to advise that
some locations have tax added. Link to DO estimated tax rates
included. Example of 20% added VAT described.
At Step 5, adds advice that DO customers
could pick lower price plans with 2 and 1 vCPUs if their
planned use of Hubs is for small groups. Adds "in the"
to the wording in Step 13, g.

Why: We've had valued Hubs community members run into
the unexpected price per month increased by VAT. This wording
advises of that possibility. It mentions the UK example
for 2 reasons:
1) Hubs community members in the UK is not
unusual at all, so we want to address their concerns.
2) It explains that the $38/month expectation would actually
be $46/month if the taxes are 20%. So this makes the added
tax example more concrete.

Fore the RAM additions, this was the result of group testing
that was reported in Discord within the past year.

Typo was merely spotted in the normal course of business.

Note: A PR for changes to the FAQ page will parallel this PR
and both relevant PR numbers will be added in GitHub later.