Add taxes and RAM Node plan options to DO section
What: Adds wording about taxes to the 'Why do the instructions say
US$38 per month minimum?' question. Adds new FAQ
'Can I select a lower price, smaller vCPU Node plan
with DigitalOcean?'' Small removal of wording that
said that the $24/month was DO's lowest price plan,
as these changes conflict that now.

Why: Both suggested changes are in the Beginner's Guide and need
to be reflected in the FAQs. Added wording is almost exactly
the same.

Note: See PR #237 for the changes to the Beginner's Guide.